chore: establish architectural quality gates
