+++ .github/workflows/server-ci.yml
@@ -70,6 +70,7 @@ jobs:
           cache: pnpm
       - run: pnpm install --frozen-lockfile
       - run: pnpm version:check
+      - run: pnpm check:architecture
       # Workspace packages publish their declarations from dist, so build the
       # dependency graph before checking downstream consumers.
       - run: pnpm build
@@ -392,6 +393,8 @@ jobs:
       - run: pnpm exec playwright install --with-deps chromium
       - run: pnpm build
       - run: cp deploy/docker/Cargo.lock.hosted-provider Cargo.lock
+      - run: cargo fmt --all --check
+      - run: cargo clippy --locked --workspace --all-targets -- -D warnings
       # Compile and test every Rust crate against the exact SDK revision used
       # by the production provider image. This keeps path-only developer
       # dependencies from making a clean checkout unreproducible.
