+++ config/architecture-budgets.json
@@ -1,0 +1,22 @@
+{
+  "productionFileMaxLines": 1000,
+  "legacyFileLineBudgets": {
+    "apps/desktop/src/main/main.ts": 1022,
+    "apps/desktop/src/renderer/main.tsx": 1714,
+    "apps/portal/src/main.tsx": 2251,
+    "crates/connect-agent/src/mirrors.rs": 1594,
+    "crates/connect-agent/src/server.rs": 1967,
+    "crates/connect-cli/src/lib.rs": 1815,
+    "crates/connect-core/src/local_sync.rs": 1209,
+    "crates/connect-core/src/registry.rs": 5324,
+    "crates/connect-hosted-provider/src/provider.rs": 6550,
+    "crates/connect-mirror/src/lib.rs": 1915,
+    "crates/connect-protocol/src/lib.rs": 1589,
+    "packages/client/src/index.ts": 4168,
+    "packages/sync/src/adoption.ts": 1127,
+    "packages/sync/src/index.ts": 1116,
+    "packages/sync/src/mirror.ts": 1255,
+    "services/server/src/app.ts": 8738,
+    "services/server/src/legacy-baseline.ts": 1114
+  }
+}
