+++ crates/connect-hosted-provider/src/provider/operation_context.rs
@@ -1,0 +1,13 @@
+use serde_json::Value;
+use uuid::Uuid;
+
+use super::Replica;
+
+pub(super) struct RecordOperationContext<'a> {
+    pub(super) collection_id: Uuid,
+    pub(super) token: &'a str,
+    pub(super) replica: &'a Replica,
+    pub(super) operation: &'a str,
+    pub(super) request_id: Uuid,
+    pub(super) request_input: &'a Value,
+}
